requirements.txt
# requirements.txt
Flask==2.3.3
werkzeug==2.3.8
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Caching
python-dotenv
requests
gunicorn
cryptography
email-validator
APScheduler==3.10.4
websockets>=12.0
pyotp>=2.9.0
qrcode[pil]>=7.4.0
aemo-to-tariff>=0.7.8
